+++ lua/config/autocmds.lua
@@ -19,17 +19,25 @@ vim.api.nvim_create_autocmd("BufEnter", { pattern = "*.md",
   callback = function() pwk.attach_markdown(0) end })
 vim.api.nvim_create_autocmd("BufEnter", { pattern = { "package.json" },
   callback = function() pwk.attach_npm(0) end })
-vim.api.nvim_create_autocmd("FileType",
-  { pattern = "*",
-    callback = function()
-      if EcoVim.plugins.zen.enabled and vim.bo.filetype ~= "alpha" then
-        pwk.attach_zen(0)
-      end
-    end
-  })
 vim.api.nvim_create_autocmd("BufEnter", { pattern = { "*test.js", "*test.ts", "*test.tsx", "*spec.ts", "*spec.tsx" },
   callback = function() pwk.attach_jest(0) end })
 vim.api.nvim_create_autocmd("FileType", { pattern = "spectre_panel",
   callback = function() pwk.attach_spectre(0) end })
 vim.api.nvim_create_autocmd("FileType", { pattern = "NvimTree",
   callback = function() pwk.attach_nvim_tree(0) end })
+
+-- File Type Plugin Lazy Loading
+local lazy = vim.api.nvim_create_augroup("lazy", {})
+vim.api.nvim_create_autocmd(
+  "UIEnter",
+  {
+    group = lazy,
+    pattern = { "*.test.ts", "*.test.tsx", "*.spec.ts", "*.spec.tsx", "*.test.js", "*.spec.js" },
+    callback = function()
+      print("lazy loading lua plugins")
+      require "lazy.loader".load({
+        plugins = { "neotest", }
+      })
+    end
+  }
+)