+++ lua/plugins/bufferline.lua
@@ -4,28 +4,19 @@ return {
     event = "VeryLazy",
     dependencies = {
       "nvim-tree/nvim-web-devicons",
-      {
-        "echasnovski/mini.bufremove",
-        version = "*",
-        config = function()
-          require("mini.bufremove").setup({
-            silent = true,
-          })
-        end,
-      },
     },
     config = function()
       local bufferline = require("bufferline")
       bufferline.setup {
         options = {
           close_command = function(n)
-            require("mini.bufremove").delete(n, false)
+            Snacks.bufdelete(n)
           end,
           right_mouse_command = function(n)
-            require("mini.bufremove").delete(n, false)
+            Snacks.bufdelete(n)
           end,
           show_buffer_close_icons = false,
-          separator_style = { "|", "|" },
+          separator_style = { "", "" },
           always_show_bufferline = true,
           style_preset = bufferline.style_preset.no_italic,
           numbers = function(opts)
@@ -39,26 +30,17 @@ return {
           end,
           offsets = {
             {
-              filetype = "NvimTree",
-              text = "File Explorer",
+              filetype = "snacks_layout_box",
+              text = "",
               highlight = "EcovimNvimTreeTitle",
               text_align = "center",
-              separator = true,
+              separator = false,
             },
           },
         },
       }
     end,
     keys = {
-      { "<Space>1",    "<cmd>BufferLineGoToBuffer 1<CR>" },
-      { "<Space>2",    "<cmd>BufferLineGoToBuffer 2<CR>" },
-      { "<Space>3",    "<cmd>BufferLineGoToBuffer 3<CR>" },
-      { "<Space>4",    "<cmd>BufferLineGoToBuffer 4<CR>" },
-      { "<Space>5",    "<cmd>BufferLineGoToBuffer 5<CR>" },
-      { "<Space>6",    "<cmd>BufferLineGoToBuffer 6<CR>" },
-      { "<Space>7",    "<cmd>BufferLineGoToBuffer 7<CR>" },
-      { "<Space>8",    "<cmd>BufferLineGoToBuffer 8<CR>" },
-      { "<Space>9",    "<cmd>BufferLineGoToBuffer 9<CR>" },
       { "<A-1>",       "<cmd>BufferLineGoToBuffer 1<CR>" },
       { "<A-2>",       "<cmd>BufferLineGoToBuffer 2<CR>" },
       { "<A-3>",       "<cmd>BufferLineGoToBuffer 3<CR>" },